+++ ServletDemo/src/com/itacademy/bobkevich/servlet/util/ResourceCategory.java
@@ -5,7 +5,7 @@ import lombok.experimental.UtilityClass;
 @UtilityClass
 public class ResourceCategory {
 
-    public static final Integer WORK = 1;
-    public static final Integer STUDY = 2;
-    public static final Integer ENTERTAINMENT = 3;
+    public static final Long WORK = 1L;
+    public static final Long STUDY = 2L;
+    public static final Long ENTERTAINMENT = 3L;
 }
